+++ app/services/vacancies/import/shared.rb
@@ -1,0 +1,9 @@
+module Vacancies::Import
+  module Shared
+    def vacancy_listed_at_excluded_school_type?(schools)
+      return false if schools.none?
+
+      (schools.map(&:detailed_school_type) & School::EXCLUDED_DETAILED_SCHOOL_TYPES).present?
+    end
+  end
+end
