+++ app/services/vacancies/import/sources/my_new_term.rb
@@ -1,10 +1,10 @@
 require "httparty"
 
-class VacancySource::Source::MyNewTerm
+class Vacancies::Import::Sources::MyNewTerm
   include HTTParty
   include Enumerable
-  include VacancySource::Parser
-  include VacancySource::Shared
+  include Vacancies::Import::Parser
+  include Vacancies::Import::Shared
 
   BASE_URI = ENV.fetch("VACANCY_SOURCE_MY_NEW_TERM_FEED_URL").freeze
   API_KEY = ENV.fetch("VACANCY_SOURCE_MY_NEW_TERM_API_KEY").freeze
